requirements.txt
aiogram~=2.18
aioredis<2.0
environs~=9.0
Pillow